Added WavPack export
