+++ src/au3wrap/CMakeLists.txt
@@ -579,6 +579,9 @@ set(AU3_SRC
     ${AU3_MODULES}/import-export/mod-mp3/ExportMP3.cpp
     ${AU3_MODULES}/import-export/mod-mp3/ExportMP3.h
     ${AU3_MODULES}/import-export/mod-mp3/MP3.cpp
+
+    ${AU3_MODULES}/import-export/mod-wavpack/ExportWavPack.cpp
+    ${AU3_MODULES}/import-export/mod-wavpack/ExportWavPack.h
 )
 
 if (AU_MODULE_EFFECTS_AUDIO_UNIT)
